core/requirements.txt
Pillow>=7.1.1
numpy>=1.18.2
zerorpc>=0.6.3
blend-modes>=2.1.0